menu-driven music playlist manager program
Create a menu-driven music playlist manager program using a singly linked list to manage your
favorite songs. Each song in the playlist will have a name.
The program should offer the following options:
• Add a Song: When the user adds a song, the program should ask the name of the song,
after that the user will be prompted to specify its "favorite level." They can choose from
the following options:
1. (Top Favorite): The song will be inserted at the beginning of the playlist.
2. (Lease Favorite): The song will be inserted at the end of the playlist.
3. Somewhere between (let me specify the number): If the user selects a custom
favorite level (e.g., 2 to 4), the program will display valid priority levels based
on the current number of songs in the playlist (e.g., 2 to 5 for a playlist of 5
songs). The song will be inserted at the specified custom priority position.
• Delete a Song: Users can remove a song from the playlist by entering its name/ position
(depends on you).
• Search for a Song: Users can search for a song by its name and display the name and
level of how much favorite it is.
• Update a song name: Thie function will take the song number, and the new name, and
update the song name on that position.
• Print Playlist: This option will display the entire playlist with song names.
• Play Song: Users can start playing songs from the beginning of the playlist. When the
user clicks on Play Song, play the first song in the list. After that the user can select the
option of play next song, so the program will play the next song from the list (just a
simulation by displaying the song name.
• Exit: Exit the program.
Note: After each insertion, update, or removal operation, print the playlist. Your program
should have all the checks for invalid input. The update song option is not showed in the
sample output, but you must implement it in your code, and display the menu accordingly.